backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
slowapi==0.1.9

# AI and ML
langchain-anthropic
langchain-core
langchain-community
langgraph
anthropic>=0.16.0

# Data processing
pydantic
pydantic-settings
PyPDF2==3.0.1

# Caching and storage
redis==5.0.1

# Observability
langfuse==3.0.7

# Utilities
python-dotenv==1.0.0